Extract engine audio extraction operation
extract_audio is a compact path-returning operation with format validation and codec selection. This moves it into engine_extract_audio.py while re-exporting extract_audio from mcp_video.engine for compatibility.

Constraint: Preserve valid formats, codec mapping, bitrate behavior, output naming, and return type.

Rejected: Move normalize_audio with extract_audio | normalize_audio is a different analysis/filter operation and should be reviewed separately.

Confidence: high

Scope-risk: narrow

Directive: Keep audio operations split by behavior rather than creating a broad bucket module.

Tested: ruff check mcp_video/engine.py mcp_video/engine_extract_audio.py

Tested: ruff format --check mcp_video/engine.py mcp_video/engine_extract_audio.py

Tested: /opt/homebrew/bin/python3 extract_audio re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_client.py::TestClientExtractAudio tests/test_cli.py::TestCLIExtractAudio tests/test_server.py::TestVideoExtractAudioTool tests/test_e2e.py::TestAudioExtractionWorkflow -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
